bot_backend/requirements.txt
Flask
openai
twilio
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client